Move the join code label to its own component.
