+++ multiplayer/src/components/GamePage.tsx
@@ -1,9 +1,7 @@
-import { faCopy } from "@fortawesome/free-regular-svg-icons";
-import { faCheck } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { useContext, useEffect, useState } from "react";
-import { AppStateContext, dispatch } from "../state/AppStateContext";
+import { useContext } from "react";
+import { AppStateContext } from "../state/AppStateContext";
 import ArcadeSimulator from "./ArcadeSimulator";
+import JoinCodeLabel from "./JoinCodeLabel";
 import Presence from "./Presence";
 import Reactions from "./Reactions";
 import ToggleMuteButton from "./ToggleMuteButton";
@@ -14,27 +12,6 @@ export default function Render(props: GamePageProps) {
     const { state } = useContext(AppStateContext);
     const { appMode } = state;
     const { netMode } = appMode;
-    const [copySuccessful, setCopySuccessful] = useState(false);
-    const copyTimeoutMs = 1000;
-
-    const copyJoinCode = async () => {
-        pxt.tickEvent("mp.copyjoincode");
-        if (state.gameState?.joinCode) {
-            navigator.clipboard.writeText(state.gameState?.joinCode);
-            setCopySuccessful(true);
-        }
-    };
-
-    useEffect(() => {
-        if (copySuccessful) {
-            let resetCopyTimer = setTimeout(() => {
-                setCopySuccessful(false);
-            }, copyTimeoutMs);
-            return () => {
-                clearTimeout(resetCopyTimer);
-            };
-        }
-    }, [copySuccessful]);
 
     return (
         <div>
@@ -48,36 +25,7 @@ export default function Render(props: GamePageProps) {
                     {state.playerSlot && <ArcadeSimulator />}
                     <div className="tw-flex tw-flex-row tw-w-full tw-items-center tw-justify-between tw-mt-1">
                         <ToggleMuteButton />
-                        <div className="tw-justify-self-center">
-                            {state.gameState?.joinCode && (
-                                <div>
-                                    {state.gameState?.joinCode &&
-                                        lf(
-                                            "Join Code: {0}",
-                                            state.gameState?.joinCode
-                                        )}
-                                    <button
-                                        onClick={copyJoinCode}
-                                        title={lf("Copy Join Code")}
-                                    >
-                                        <div className="tw-text-sm tw-ml-1">
-                                            {!copySuccessful && (
-                                                <FontAwesomeIcon
-                                                    icon={faCopy}
-                                                    className="hover:tw-scale-105 tw-mb-[0.1rem]"
-                                                />
-                                            )}
-                                            {copySuccessful && (
-                                                <FontAwesomeIcon
-                                                    icon={faCheck}
-                                                    className="tw-text-green-600 tw-mb-[0.1rem]"
-                                                />
-                                            )}
-                                        </div>
-                                    </button>
-                                </div>
-                            )}
-                        </div>
+                        <JoinCodeLabel />
                         <div>{lf("Keyboard Controls")}</div>
                     </div>
                     <div className="tw-flex tw-flex-row tw-space-x-2 tw-items-center tw-align-middle tw-justify-center tw-mt-3">
